Separo IndicatorMetricsCalculator en otro archivo
