+++ api_management/apps/analytics/models.py
@@ -1,10 +1,9 @@
 import abc
 import hashlib
 import re
+import uuid
 from urllib.parse import parse_qsl, urlparse
 
-import uuid
-from datetime import date
 import requests
 from dateutil import relativedelta
 from django.conf import settings
@@ -214,63 +213,3 @@ class IndicatorMetricsRow(models.Model):
     all_mobile = models.IntegerField(blank=True, null=True)
     all_not_mobile = models.IntegerField(blank=True, null=True)
     total_users = models.IntegerField(blank=True, null=True)
-
-
-class IndicatorMetricsCalculator:
-
-    def __init__(self, api_name):
-        self.api_name = api_name
-
-    def is_mobile(self, user_agent):
-        return "Mobile" in user_agent or \
-               "Android" in user_agent or \
-               "iPhone" in user_agent or \
-               "Slackbot" in user_agent
-
-    def indicator_row_content(self, queries):
-        unique_session_ids = set()
-        total_mobile = 0
-        total_not_mobile = 0
-
-        for query in queries:
-            unique_session_ids.add(generate_api_session_id(query))
-
-            if self.is_mobile(query.user_agent):
-                total_mobile = total_mobile + 1
-            else:
-                total_not_mobile = total_not_mobile + 1
-
-        return {'total': total_mobile + total_not_mobile,
-                'total_mobile': total_mobile,
-                'total_not_mobile': total_not_mobile,
-                'total_unique_users': len(unique_session_ids)}
-
-    def first_query_time(self):
-        query_time = Query.objects.first().start_time
-        last_row = IndicatorMetricsRow.objects.filter(api_name=self.api_name).last()
-        if last_row is not None:
-            query_time = last_row.date
-
-        return query_time
-
-    def calculate(self):
-        query_time = self.first_query_time()
-
-        while query_time.date() < date.today():
-            queries = self.all_queries(query_time)
-            total_counts = self.indicator_row_content(queries)
-
-            indicator_row = IndicatorMetricsRow(api_name=self.api_name)
-            indicator_row.date = query_time.date()
-            indicator_row.all_queries = total_counts.get('total')
-            indicator_row.all_mobile = total_counts.get('total_mobile')
-            indicator_row.all_not_mobile = total_counts.get('total_not_mobile')
-            indicator_row.total_users = total_counts.get('total_unique_users')
-            indicator_row.save()
-            query_time = next_day_of(query_time)
-
-    def all_queries(self, query_time):
-        return Query.objects.filter(api_data__name=self.api_name,
-                                    start_time__gte=query_time,
-                                    start_time__lt=next_day_of(query_time))\
-                            .exclude(request_method='OPTIONS')
